REFACTOR: Changes checkUserLoggedIn to checkUserAuthorization
